refactor(navigation): extract policies and providers
