+++ apps/backend/app/domains/navigation/application/router.py
@@ -3,18 +3,14 @@ from __future__ import annotations
 import copy
 import logging
 import math
-import random
 import time
-from abc import ABC, abstractmethod
 from collections import Counter, defaultdict, deque
 from collections.abc import Sequence
 from dataclasses import dataclass
 from enum import Enum
 from typing import TYPE_CHECKING
-from uuid import UUID
 
 from sqlalchemy.ext.asyncio import AsyncSession
-from sqlalchemy.future import select
 
 from app.core.preview import PreviewContext  # isort: skip
 from app.core.log_events import no_route, transition_finish, transition_start
@@ -30,13 +26,9 @@ from app.domains.telemetry.application.transition_metrics_facade import (
     transition_metrics,
 )
 
+from .policies import Policy
+
 if TYPE_CHECKING:  # pragma: no cover - used for type hints only
-    from app.domains.navigation.application.compass_service import (
-        CompassService,
-    )
-    from app.domains.navigation.application.transitions_service import (
-        TransitionsService,
-    )
     from app.domains.nodes.infrastructure.models.node import Node
     from app.domains.users.infrastructure.models.user import User
 
@@ -49,128 +41,6 @@ def _compute_entropy(tags: Sequence[str]) -> float:
     counts = Counter(tags)
     total = sum(counts.values())
     return -sum((c / total) * math.log(c / total) for c in counts.values())
-
-
-class TransitionProvider(ABC):
-    """Interface for objects that return possible transitions from a node."""
-
-    @abstractmethod
-    async def get_transitions(
-        self,
-        db: AsyncSession,
-        node: Node,
-        user: User | None,
-        workspace_id: UUID,
-        preview: PreviewContext | None = None,
-    ) -> Sequence[Node]:
-        """Return candidate nodes for transition."""
-
-
-class Policy(ABC):
-    """Policy defines how to pick next transition from provider's output."""
-
-    name: str
-
-    def __init__(self, provider: TransitionProvider) -> None:
-        self.provider = provider
-
-    @abstractmethod
-    async def choose(
-        self,
-        db: AsyncSession,
-        node: Node,
-        user: User | None,
-        history: deque[str],
-        repeat_filter: RepeatFilter,
-        preview: PreviewContext | None = None,
-    ) -> tuple[Node | None, TransitionTrace]:
-        """Return next node and trace info or ``None`` if not applicable."""
-
-
-class ManualTransitionsProvider(TransitionProvider):
-    def __init__(self, service: TransitionsService | None = None) -> None:
-        if service is None:
-            from app.domains.navigation.application.transitions_service import (
-                TransitionsService,
-            )
-
-            service = TransitionsService()
-        self._service = service
-
-    async def get_transitions(
-        self,
-        db: AsyncSession,
-        node: Node,
-        user: User | None,
-        workspace_id: UUID,
-        preview: PreviewContext | None = None,
-    ) -> Sequence[Node]:
-        transitions = await self._service.get_transitions(
-            db, node, user, workspace_id, preview=preview
-        )
-        return [t.to_node for t in transitions]
-
-
-class CompassProvider(TransitionProvider):
-    def __init__(self, service: CompassService | None = None, limit: int = 5) -> None:
-        if service is None:
-            from app.domains.navigation.application.compass_service import (
-                CompassService,
-            )
-
-            service = CompassService()
-        self._service = service
-        self._limit = limit
-
-    async def get_transitions(
-        self,
-        db: AsyncSession,
-        node: Node,
-        user: User | None,
-        workspace_id: UUID,
-        preview: PreviewContext | None = None,
-    ) -> Sequence[Node]:
-        nodes = await self._service.get_compass_nodes(
-            db, node, user, self._limit, preview=preview
-        )
-        return [n for n in nodes if n.workspace_id == workspace_id]
-
-
-class RandomProvider(TransitionProvider):
-    """Provide random candidates using own RNG for reproducibility."""
-
-    def __init__(self) -> None:
-        self._rnd = random.Random()
-
-    def set_seed(self, seed: int | None) -> None:
-        self._rnd.seed(seed)
-
-    async def get_transitions(
-        self,
-        db: AsyncSession,
-        node: Node,
-        user: User | None,
-        workspace_id: UUID,
-        preview: PreviewContext | None = None,
-    ) -> Sequence[Node]:
-        from app.core.pagination import scope_by_workspace
-        from app.domains.navigation.application.access_policy import has_access_async
-        from app.domains.nodes.infrastructure.models.node import Node
-
-        query = select(Node).where(
-            Node.is_visible,
-            Node.is_public,
-            Node.is_recommendable,
-            Node.id != node.id,
-            Node.workspace_id == workspace_id,
-        )
-        query = scope_by_workspace(query, workspace_id)
-        result = await db.execute(query)
-        nodes: list[Node] = result.scalars().all()
-        nodes = [n for n in nodes if await has_access_async(n, user, preview)]
-        if not nodes:
-            return []
-        return [self._rnd.choice(nodes)]
 
 
 class RepeatFilter:
@@ -235,106 +105,6 @@ class RepeatFilter:
                     continue
             allowed.append(n)
         return allowed, filtered
-
-
-class ManualPolicy(Policy):
-    name = "manual"
-
-    async def choose(
-        self,
-        db: AsyncSession,
-        node: Node,
-        user: User | None,
-        history: deque[str],
-        repeat_filter: RepeatFilter,
-        preview: PreviewContext | None = None,
-    ) -> tuple[Node | None, TransitionTrace]:
-        try:
-            candidates = await self.provider.get_transitions(
-                db, node, user, node.workspace_id, preview=preview
-            )
-        except TypeError:
-            candidates = await self.provider.get_transitions(
-                db, node, user, node.workspace_id
-            )
-        candidate_slugs = [n.slug for n in candidates]
-        filtered = [n.slug for n in candidates if n.slug in history]
-        candidates = [n for n in candidates if n.slug not in history]
-        candidates, filt2 = repeat_filter.filter(candidates)
-        filtered.extend(filt2)
-        for n in candidates:
-            return n, TransitionTrace(candidate_slugs, filtered, {}, n.slug)
-        return None, TransitionTrace(candidate_slugs, filtered, {}, None)
-
-
-class CompassPolicy(Policy):
-    name = "compass"
-
-    async def choose(
-        self,
-        db: AsyncSession,
-        node: Node,
-        user: User | None,
-        history: deque[str],
-        repeat_filter: RepeatFilter,
-        preview: PreviewContext | None = None,
-    ) -> tuple[Node | None, TransitionTrace]:
-        try:
-            candidates = await self.provider.get_transitions(
-                db, node, user, node.workspace_id, preview=preview
-            )
-        except TypeError:
-            candidates = await self.provider.get_transitions(
-                db, node, user, node.workspace_id
-            )
-        candidate_slugs = [n.slug for n in candidates]
-        filtered = [n.slug for n in candidates if n.slug in history]
-        candidates = [n for n in candidates if n.slug not in history]
-        candidates, filt2 = repeat_filter.filter(candidates)
-        filtered.extend(filt2)
-        for n in candidates:
-            return n, TransitionTrace(candidate_slugs, filtered, {}, n.slug)
-        return None, TransitionTrace(candidate_slugs, filtered, {}, None)
-
-
-class RandomPolicy(Policy):
-    name = "random"
-
-    def __init__(self, provider: TransitionProvider) -> None:
-        super().__init__(provider)
-        self._rnd = random.Random()
-
-    def set_seed(self, seed: int | None) -> None:
-        self._rnd.seed(seed)
-        if hasattr(self.provider, "set_seed"):
-            self.provider.set_seed(seed)
-
-    async def choose(
-        self,
-        db: AsyncSession,
-        node: Node,
-        user: User | None,
-        history: deque[str],
-        repeat_filter: RepeatFilter,
-        preview: PreviewContext | None = None,
-    ) -> tuple[Node | None, TransitionTrace]:
-        try:
-            candidates = await self.provider.get_transitions(
-                db, node, user, node.workspace_id, preview=preview
-            )
-        except TypeError:
-            candidates = await self.provider.get_transitions(
-                db, node, user, node.workspace_id
-            )
-        candidate_slugs = [n.slug for n in candidates]
-        filtered = [n.slug for n in candidates if n.slug in history]
-        candidates = [n for n in candidates if n.slug not in history]
-        candidates, filt2 = repeat_filter.filter(candidates)
-        filtered.extend(filt2)
-        if not candidates:
-            return None, TransitionTrace(candidate_slugs, filtered, {}, None)
-        chosen = self._rnd.choice(candidates)
-        return chosen, TransitionTrace(candidate_slugs, filtered, {}, chosen.slug)
 
 
 @dataclass
